requirements.txt
# Data Handling
numpy
pandas
scipy
joblib

# Machine Learning & Modeling
scikit-learn
statsmodels
xgboost

# API & Data Fetching
requests
yfinance
ccxt

# Database
sqlalchemy
psycopg2

# Notifications
twilio
python-dotenv

# Scheduling & Automation
schedule
APScheduler

# Visualization
matplotlib
seaborn
plotly
mplfinance

# Git & Version Control
gitpython
